requirements.txt
fastapi
uvicorn
chromadb
pydantic
python-multipart
typing
numpy
langchain
langchain-community
openai
python-dotenv
requests
beautifulsoup4
selenium
tavily-python
langchain-openai
nest_asyncio
xmltodict

motor
pymongo
langgraph
langchain-deepseek
xmltodict
aiohttp
python-jose[cryptography]
passlib
bcrypt
motor
pydantic-settings
langchain-core
tiktoken
duckduckgo-search
sentence-transformers
langchain-chroma
spacy

